feat: Ts&Cs payment and summary components split out, tests applied
