+++ support-frontend/assets/pages/supporter-plus-landing/components/TsAndCsRenewal.tsx
@@ -1,0 +1,22 @@
+import type { ContributionType } from 'helpers/contributions';
+import {
+	getDateWithOrdinal,
+	getLongMonth,
+} from 'helpers/utilities/dateFormatting';
+
+export default function TsAndCsRenewal({
+	contributionType,
+}: {
+	contributionType: ContributionType;
+}): JSX.Element {
+	const today = new Date();
+	if (contributionType === 'ANNUAL') {
+		return (
+			<>
+				on the {getDateWithOrdinal(today)} day of {getLongMonth(today)} every{' '}
+				year
+			</>
+		);
+	}
+	return <>on the {getDateWithOrdinal(today)} day of every month</>;
+}
